+++ main_page/libs/dicomlib.py
@@ -1,4 +1,4 @@
-import pydicom, datetime, logging
+import pydicom, datetime, logging, csv
 from pydicom.values import convert_SQ, convert_string
 from pydicom.dataset import Dataset
 from pydicom.sequence import Sequence
@@ -296,6 +296,10 @@ def try_add_sample_sequence(ds: Type[Dataset], sample_seq: List[Tuple[datetime.d
     logger.info('Removing Seqence')
     del ds[0x00231020]
 
+def try_add_dicom_history(ds: Type[Dataset], dicom_history):
+  if dicom_history:
+    ds.clearancehistory = Sequence(dicom_history)
+
 
 def try_add_pixeldata(ds: Type[Dataset], pixeldata: bytes) -> None:
   """
@@ -343,6 +347,7 @@ def fill_dicom(ds,
     clearance_norm      = None,
     cpr                 = None,
     department          = None,
+    dicom_history       = None,
     exam_status         = None,
     gender              = None,
     gfr                 = None,
@@ -382,6 +387,7 @@ def fill_dicom(ds,
     clearance_norm      : float, Clearance Value Normalized to 1.73m²
     cpr                 : string, CPR number
     department          :
+    dicom_history       : list[class:pydicom:dataset], contains the history of the patient
     exam_status         : int, status of the exam, e.g. to be reviewed, ready to send to packs, etc.
     gender              :
     gfr                 : string, either 'Normal', 'Moderat Nedsat', 'Nedsat', 'Stærkt nedsat' 
@@ -477,9 +483,83 @@ def fill_dicom(ds,
     try_add_gender: [gender],
     try_add_pixeldata: [pixeldata],
     # ### PRIVATE TAGS START ###
-    try_add_sample_sequence: [sample_seq]
+    try_add_sample_sequence: [sample_seq],
+    try_add_dicom_history: [dicom_history]
   }
 
   for try_func, args in custom_try_adds.items():
     # Args is 'unpacked' to allow for functions with none or multiple required arguments
     try_func(ds, *args)
+
+def export_dicom(ds, file_path):
+  """
+    converts a dicom file to csv file and saves it at 'file_path'
+    Note that this csv file doesn't contain any patient history
+
+
+    args:
+      ds: Pydicom dataset, data to be saved
+      file_path: str, the file destination for the csv
+    Throws:
+      Attribute Error: if the ds do not have the following tags
+        [0x00100010]
+
+    
+  """
+
+  with open(file_path, mode='w', newline='') as csv_file:
+    #init Writeer
+    csv_writer = csv.writer(
+      csv_file,
+      delimiter=',',
+      quotechar='"'     
+    )
+    #Data to put into file
+    try:
+      #Format of the list is tuple, with (header, data)
+      #Keep it to this format for easy commenting out / addition
+      print(f'{type(ds.PatientName)}, {ds.PatientName}')
+      column_list = [
+        ('CPR',                       ds.PatientID),
+        ('Navn',                      formatting.person_name_to_name(str(ds.PatientName))),
+        ('Study Date',                ds.StudyDate),
+        ('Height',                    ds.PatientSize * 100),
+        ('Weight',                    ds.PatientWeight),
+        ('Body Surface Area Method',  ds.BSAmethod),
+        ('Standard count',            ds.stdcnt),
+        ('Thining Factor',            ds.thiningfactor),
+        ('Injection Weight',          ds.injWeight),
+        ('Injection Time',            ds.injTime), 
+        ('Clearance',                 ds.clearance),
+        ('Clearance Normalized',      ds.normClear)
+      ]
+
+      #Get rows 
+      header_row = [ column[0] for column in column_list]
+      data_row   = [ column[1] for column in column_list] 
+      #Write Data
+      csv_writer.writerow(header_row)
+      csv_writer.writerow(data_row)
+    except AttributeError as E: #Missing a tag
+      logger.info(f'Error in writing to CSV: {E}')
+      return 'Incomplete dicom 1'
+
+    sample_header = [
+      'Sample Value',
+      'Sample time'
+    ]
+    csv_writer.writerow(sample_header)
+
+    try:
+      for sample_ds in ds.ClearTest:
+        sample = [
+          sample_ds[0x00231022].value,
+          sample_ds.SampleTime
+        ]
+        csv_writer.writerow(sample)
+
+    except AttributeError as E:
+      logger.info(f'Error in writing to CSV: {E}')
+      return 'Incomplete Dicom 2'
+  
+  return 'OK'