+++ main_page/libs/formatting.py
@@ -295,7 +295,7 @@ def reverse_format_date(reverse_Date : str, sep='') -> str:
 
     Raises:
       ValueError : On invalid String
-    """
+    """ 
   if reverse_Date == None or reverse_Date == '':
     return ''
   # Argument checking