+++ main_page/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 from django.conf.urls import (handler400, handler403, handler404, handler500)
 
 #Sooo WhY do we not just import api at this point?
-from main_page.views.api.api import UserEndpoint, HospitalEndpoint, DepartmentEndpoint, ConfigEndpoint, HandledExaminationsEndpoint, SambaBackupEndpoint, ProcedureEndpoint, ProcedureMappingsEndpoint, StudyEndpoint
+from main_page.views.api.api import UserEndpoint, HospitalEndpoint, DepartmentEndpoint, ConfigEndpoint, HandledExaminationsEndpoint, SambaBackupEndpoint, ProcedureEndpoint, ProcedureMappingsEndpoint, StudyEndpoint, CsvEndpoint
 
 import main_page.views.views as views
 
@@ -53,4 +53,6 @@ urlpatterns = [
   path('api/procedure_mapping', ProcedureMappingsEndpoint.as_view(), name='procedure_mapping'),
   path('api/procedure_mapping/<int:obj_id>', ProcedureMappingsEndpoint.as_view(), name='procedure_mapping'),
   path('api/study/<str:ris_nr>', StudyEndpoint.as_view(), name='study'),
+  path('api/csv/<str:accessionnumber>', CsvEndpoint.as_view(), name='csv'),
+  #path('api/delete_csv/<str:accessionnumber>') #Path incomplete
 ]